+++ src/Model/Product.php
@@ -2,7 +2,7 @@
 /**
  * This file is part of the Billbee API package.
  *
- * Copyright 2017 - 2019 by Billbee GmbH
+ * Copyright 2017 - 2020 by Billbee GmbH
  *
  * For the full copyright and license information, please read the LICENSE
  * file that was distributed with this source code.
@@ -312,8 +312,8 @@ class Product
     public $unitsPerItem = 1.0;
 
     /**
-     * @var int
-     * @DataField(name="SoldAmount", type="int")
+     * @var float
+     * @DataField(name="SoldAmount", type="float")
      */
     public $soldAmount = 0;